+++ app/src/main/java/com/scouts/kitchenplaner/ui/view/createProject/ProjectCreation.kt
@@ -30,18 +30,24 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 
 @Composable
-fun ProjectCreation(onNavigateToInvitePeople: (Int) -> Unit){
+fun ProjectCreation(onNavigateToInvitePeople: (Long) -> Unit) {
     var projectId by remember { mutableStateOf(0f) }
 
     Column {
         Text("Here it is possible to create a new Project")
-        Text(text= "available Links to other sides are: ")
-        
+        Text(text = "available Links to other sides are: ")
+
         Row {
             Text(text = "inviteToProject")
-            Slider(modifier = Modifier.fillMaxWidth(0.3f),value = projectId, onValueChange = {projectId = it }, valueRange = 1f..10f, steps = 1 )
-            Button(onClick =  {onNavigateToInvitePeople(projectId.toInt())}) {
-                
+            Slider(
+                modifier = Modifier.fillMaxWidth(0.3f),
+                value = projectId,
+                onValueChange = { projectId = it },
+                valueRange = 1f..10f,
+                steps = 1
+            )
+            Button(onClick = { onNavigateToInvitePeople(projectId.toLong()) }) {
+
             }
         }
