+++ app/src/main/java/com/scouts/kitchenplaner/ui/view/recipeDetails/RecipeDetails.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.graphics.Color
 
 
 @Composable
-fun recipeDetails(recipeID: Int) {
+fun recipeDetails(recipeID: Long) {
 
     Column {
         Text("On this screen you can see a detailed recipe")
